+++ src/Imprint.js
@@ -1,0 +1,28 @@
+import React from 'react';
+
+function Imprint() {
+  return (
+    <>
+      <h2>Legal Notice</h2>
+      <p>
+        Jan Kohlbach
+        <br />
+        Reinhold-Frank-Straße 51
+        <br />
+        76133 Karlsruhe
+      </p>
+      <p>
+        <span>Fon: </span>
+        <a href="tel:+4917643675791">+49 176 43675791</a>
+        <br />
+        <span>E-Mail: </span>
+        <a href="mailto:contact@b-reminded.com">contact@b-reminded.com</a>
+        <br />
+        <span>Web: </span>
+        <a href="https://jankohlbach.com">jankohlbach.com</a>
+      </p>
+    </>
+  );
+}
+
+export default Imprint;